# src/main/docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PaymentHttp.CreateCheckoutOutput": {
				"properties": {
					"status": {
						"type": "string"
					},
					"paymentId": {
						"type": "string"
					},
					"orderId": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"paymentId",
					"orderId"
				],
				"type": "object"
			},
			"Payment.Order": {
				"properties": {
					"totalValue": {
						"type": "number",
						"format": "double"
					},
					"clientId": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"orderId": {
						"type": "string"
					}
				},
				"required": [
					"totalValue",
					"status",
					"orderId"
				],
				"type": "object"
			},
			"PaymentHttp.CreateCheckoutInput": {
				"properties": {
					"paymentMethod": {
						"type": "string"
					},
					"order": {
						"$ref": "#/components/schemas/Payment.Order",
						"description": "paymentMethod aceita somente 'PIX'"
					}
				},
				"required": [
					"paymentMethod",
					"order"
				],
				"type": "object"
			},
			"Payment.FindPaymentOutput": {
				"properties": {
					"orderId": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"expirationDate": {
						"type": "string",
						"format": "date-time"
					},
					"pixCode": {
						"type": "string"
					},
					"pixUrl": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"paymentMethod": {
						"type": "string"
					},
					"totalValue": {
						"type": "number",
						"format": "double"
					},
					"paymentId": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"orderId",
					"expirationDate",
					"pixCode",
					"pixUrl",
					"status",
					"paymentMethod",
					"totalValue",
					"paymentId",
					"id"
				],
				"type": "object"
			},
			"PaymentHttp.GetPaymentOutput": {
				"$ref": "#/components/schemas/Payment.FindPaymentOutput"
			},
			"PaymentHttp.UpdatePaymentStatusOutput": {
				"properties": {
					"paymentId": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"paymentId",
					"status"
				],
				"type": "object"
			},
			"PaymentService.PaymentWebhookInput": {
				"properties": {
					"data": {
						"properties": {
							"currency": {
								"type": "string"
							},
							"amount": {
								"type": "number",
								"format": "double"
							},
							"code": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"currency",
							"amount",
							"code",
							"id"
						],
						"type": "object"
					},
					"created_at": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"id": {
						"type": "string",
						"description": "Weebhook recebido do gateway de pagamento"
					}
				},
				"required": [
					"data",
					"created_at",
					"type",
					"id"
				],
				"type": "object"
			},
			"PaymentHttp.UpdatePaymentStatusInput": {
				"$ref": "#/components/schemas/PaymentService.PaymentWebhookInput"
			}
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Restaurante ACME API",
		"version": "1.0.0",
		"description": "Documentação da API do restaurante ACME",
		"license": {
			"name": "GPL-3.0-or-later"
		},
		"contact": {
			"name": "Mauricio Carvalho Pinheiro",
			"email": "developer.mauricio1@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/checkout": {
			"post": {
				"operationId": "CreatePayment",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentHttp.CreateCheckoutOutput"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"summary": "Rota para criação do checkout",
				"tags": [
					"Checkout"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PaymentHttp.CreateCheckoutInput"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindPayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentHttp.GetPaymentOutput"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"summary": "Rota para obter um checkout",
				"tags": [
					"Checkout"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "paymentId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "orderId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/webhook": {
			"post": {
				"operationId": "UpdatePaymentStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentHttp.UpdatePaymentStatusOutput"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"summary": "Rota para atualizar o status do pagamento",
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PaymentHttp.UpdatePaymentStatusInput"
							}
						}
					}
				}
			}
		},
		"/healthcheck": {
			"get": {
				"operationId": "GetHealthCheck",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string",
											"enum": [
												"OK"
											],
											"nullable": false
										},
										"uptime": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"message",
										"uptime"
									],
									"type": "object"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					}
				},
				"summary": "Rota para verificação da saúde do servidor",
				"tags": [
					"Health Check"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/v1/api"
		}
	]
}